Rename Example1 to MultipleComponents, add a parameter to disable the vector icons
