+++ benchmarks/multiplatform/benchmarks/src/wasmJsMain/kotlin/main.wasmJs.kt
@@ -36,13 +36,15 @@ fun mainD8(args: Array<String>) {
     println("mainD8 is intentionally doing nothing. Read the comments in main.wasmJs.kt")
 }
 
-// Using only SIMPLE mode, because VSYNC_EMULATION calls delay(...),
-// which is implemented via setTimeout on web targets.
-// setTimeout implementation is simplified in D8, making
-// the VSYNC_EMULATION mode meaningless when running in D8
 private val basicConfigForD8 = Config(
+    // Using only SIMPLE mode, because VSYNC_EMULATION calls delay(...),
+    // which is implemented via setTimeout on web targets.
+    // setTimeout implementation is simplified in D8, making
+    // the VSYNC_EMULATION mode meaningless when running in D8
     modes = setOf(Mode.SIMPLE),
-    unsupportedBenchmarks = setOf("Example1"),
+    // MultipleComponents is unsupported, because it uses vector icons. D8 doesn't provide XML parsing APIs.
+    // But there is an alternative workload called 'MultipleComponents-NoVectorGraphics'
+    unsupportedBenchmarks = setOf("MultipleComponents"),
 )
 
 @JsExport
